// src/app/events/events.interface.ts
export interface IEvent {
  event_id: number;
  event_name: string;
  competition_id: number;
  naziv?: string;
  event_start_time: '2022-09-19T07:55:59.945Z';
  event_status?: string;
  event_period?: string;
  result?: string;
  home_team?: string;
  away_team?: string;
  yellow_cards_home_team?: number;
  yellow_cards_away_team?: number;
  red_cards_home_team?: number;
  red_cards_away_team?: number;
  eventkey?: string;
}